requirements.txt
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.31.0
accelerate>=1.2.0
transformers>=4.40.0
safetensors>=0.4.0
xformers>=0.0.29
bitsandbytes>=0.45.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
tensorboard>=2.15.0
ftfy>=6.1.0
regex>=2023.0
tqdm>=4.65.0
peft>=0.13.0
